build: renamed components and file names
