+++ client/components/lib/colors-list.js
