+++ client/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './containers/app';
+import App from './containers/App';
 
 ReactDOM.render(<App />, document.querySelector('#root'));
