+++ server/public/index.html
@@ -25,7 +25,6 @@
     <link rel="stylesheet" href="/stylesheets/main-category.css" />
     <link rel="stylesheet" href="/stylesheets/brands.css" />
     <link rel="stylesheet" href="/stylesheets/product.css" />
-    <link rel="stylesheet" href="/stylesheets/color.css" />
     <link rel="stylesheet" href="/stylesheets/colors.css" />
     <link rel="stylesheet" href="/stylesheets/spinner.css" />
   </head>
